requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0

# ML & Models
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.2
tensorflow==2.14.0
torch==2.0.1
torchvision==0.15.2

# MLOps & Monitoring
mlflow==2.8.1
wandb==0.15.12
optuna==3.14.0
dvc==3.42.0

# Data Validation & Quality
great-expectations==0.17.12
pandera==0.17.1
evidently==0.4.1

# Causal Inference & Explainability
dowhy==0.10.1
shap==0.42.1
lime==0.2.0
fairlearn==0.9.0
alibi==0.9.5

# API & Serving
fastapi==0.104.1
pydantic==2.4.2
uvicorn==0.24.0
ray==2.8.1
redis==5.0.1

# Testing & Quality
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5

# Visualization & EDA
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
bokeh==3.3.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
python-dateutil==2.8.2

# Jupyter & Notebooks
jupyter==1.0.0
jupyterlab==4.0.9
ipywidgets==8.1.1
notebook==7.0.6

# Distributed Computing
pyspark==3.5.0
dask==2023.11.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
mkdocs==1.5.3

# AWS (if using S3, SageMaker, etc)
boto3==1.28.84
sagemaker==2.186.0

# GCP (if using BigQuery, Vertex AI)
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.0
